fix: 32-bit targets (widen int overflows of MaxUint32/MaxInt64 constants)
On 32-bit GOARCHs (386, arm) the platform int is 32-bit, so untyped constants
math.MaxUint32 / 0xFFFFFFFF / math.MaxInt64 overflow int in bounds checks, map
literals, and format args, breaking cross-compilation (e.g. gomobile bind for
armeabi-v7a/x86). Widen the comparisons/types so the package builds on all
targets; no behavior change on valid inputs.

+++ consensus/byron/byron.go
@@ -156,8 +156,9 @@ func NewByronConfigFromGenesis(genesis *ledgerbyron.ByronGenesis) (ByronConfig, 
 	}
 
 	// Validate protocol magic fits in uint32
-	if genesis.ProtocolConsts.ProtocolMagic < 0 || genesis.ProtocolConsts.ProtocolMagic > math.MaxUint32 {
-		return ByronConfig{}, fmt.Errorf("invalid protocol magic: %d (must be 0 to %d)", genesis.ProtocolConsts.ProtocolMagic, math.MaxUint32)
+	// int64 cast: on 32-bit targets untyped math.MaxUint32 overflows int
+	if genesis.ProtocolConsts.ProtocolMagic < 0 || int64(genesis.ProtocolConsts.ProtocolMagic) > math.MaxUint32 {
+		return ByronConfig{}, fmt.Errorf("invalid protocol magic: %d (must be 0 to %d)", genesis.ProtocolConsts.ProtocolMagic, uint32(math.MaxUint32))
 	}
 
 	// Extract genesis delegate key hashes
